+++ RestaurantApplication/src/main/java/com/restaurant/diningReview/DiningReviewRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface DiningReviewRepository extends CrudRepository<DiningReview, Integer> {
+public interface DiningReviewRepository extends CrudRepository<DiningReview, Long> {
     List<DiningReview> findDiningReviewByStatus(Status status);
     List<DiningReview> findDiningReviewByStatusAndRestaurantId(Status status, Long restaurantId);
